Support fractional credits via half-credit scale (BYOK: 0.5 / 1 / 1.5)
Internal scale = 2 (1 display credit = 2 DB units), so 0.5-credit increments
fit in the existing integer columns without schema changes beyond migration.

Migration V10:
- Multiply ai_credit_ledger credits_used and credits_limit by 2
- Multiply ai_generations credits_cost by 2 (history stays accurate)

Backend:
- AiGenerationService: creditCost() returns 2/4/6, byokCreditCost() returns 1/2/3
  BYOK display: SHORT=0.5, MEDIUM=1, LONG=1.5 credits
  Platform display: SHORT=1, MEDIUM=2, LONG=3 credits
- GenerationResult.creditsCost changed to double (raw/2.0)
- history() maps getCreditsCost() through toDisplay()
- AiCreditService: monthlyLimit() multiplied by CREDIT_SCALE; Usage fields double
  usage() divides raw ledger values by 2 before returning
- CreditUsageResponse: double fields

Frontend:
- API Keys page table: 0.5 / 1 / 1.5 credits for BYOK
- Settings shortcut: updated description

+++ backend/src/main/java/com/zyntral/modules/ai/web/dto/AiDtos.java
@@ -48,5 +48,5 @@ public final class AiDtos {
             String model
     ) {}
 
-    public record CreditUsageResponse(int limit, int used, int remaining) {}
+    public record CreditUsageResponse(double limit, double used, double remaining) {}
 }
